PR comments: Refactored PersistentConnection class into it's own file, added max_depth as a parameter that can be decided when writing pytests, and increased maxUnconfirmedMessages/Bytes in the schema that determines the consumer/producer specifications
